Create Mini_MultiLayerPerceptron_Garuda-o1.py
This is my first self-made AI, which is a model or architecture of a Multi-layer perceptron neural network technology developed in the 1950s. With 2 inputs, 2 hidden layers and 4 neurons and 1 output, it can calculate and predict the class of input x, but this model does not learn like giving a problem to a child, but every time he answers wrong, you explain why he is wrong and help him fix his method. Over time he gets better at it.